Adicionando concern, limpando arquivos desnecessarios e ajustando testes
