+++ spec/routing/api/v1/stores_routing_spec.rb
